fix: fix the nosimd tests
