+++ tests/iteration.rs
@@ -1,4 +1,5 @@
 #[cfg(feature = "default")]
+
 mod tests {
     use coitrees::{COITree, Interval};
